+++ services/frontend-v3/src/components/errorResult/errorResultView.jsx
@@ -1,14 +1,17 @@
 import React from "react";
 import { Result } from "antd";
 import PropTypes from "prop-types";
+import AppLayout from "../layouts/appLayout/AppLayout";
 
 const ErrorResultView = ({ resultProps }) => (
-  <Result
-    status={resultProps.status}
-    title={resultProps.title}
-    subTitle={resultProps.subTitle}
-    extra={resultProps.extra}
-  />
+  <AppLayout displaySearch={false}>
+    <Result
+      status={resultProps.status}
+      title={resultProps.title}
+      subTitle={resultProps.subTitle}
+      extra={resultProps.extra}
+    />
+  </AppLayout>
 );
 
 ErrorResultView.propTypes = {
